+++ paper/paper.bib
@@ -91,7 +91,8 @@
     title={ParaView: An End-User Tool for Large-Data Visualization},
     author={James P. Ahrens and Berk Geveci and C. Charles Law},
     booktitle={The Visualization Handbook},
-    year={2005}
+    year={2005},
+    doi = {10.1016/b978-012387582-2/50038-1}
 }
 
 @manual{qgis,
