### 2022-10-06 eee1031
[Paper] Fixed minor misspelling
