### 2022-12-05 Artelia/NIMPHS#14 @liberostelios
Improve installation process
### Description

As things stand, there is a lot of manual work required by the user to install the python dependencies and enough knowledge to understand  "where python is installed in Blender", as described in the documentation. Unfortunately, this is part of the inevitable complexity of using Blender. I've had a similar issue when trying to use `pyvista` for a QGIS plugin and I tried to solve it by running the necessary python commands from inside QGIS the first time (see [here](https://github.com/liberostelios/qgis-3d-toolbox/blob/main/processing/system/install_pyvista_algorithm.py)).

This might not be the most elegant or robust solution, but it works. Maybe you could try to implement something similar when Blender loads the plugin for the first time and only fall back to the manual installation if this fails.

### Links to documentation, examples or class definitions

_No response_

### Pseudocode or screenshots

_No response_
