### 2022-12-05 #4868 @liberostelios
This is a great piece of software! Very nice @Failxxx!

Everything seems very well documented, from a software engineering point of view. The documentation is concise and clear, it made me understand the tool's purpose and way of working. The dev documentation is top notch, as well. Testing is present and everything is automated so that it can be run with minimal effort.

I can only give a small suggestion for improvement with respect to the installation process (made an issue about it [here](https://github.com/Artelia/NIMPHS/issues/14)).

From my standpoint, I suggest that this gets accepted. Again, congratulations @Failxxx for the fine work! 👌
