### 2022-12-09 Artelia/NIMPHS#14 @bnase
I agree with @liberostelios . Moreover, at least for now, you should add that it _might require elevated access_ command prompt when installing the Python packages!
