### 2022-12-09 Artelia/NIMPHS#15 @bnase
Mac support
### Description

Could we have this plugin work with macFoam? Or should we state that it supports Mac through Docker (e.g. through an Ubuntu latest Canonical image?)

### Links to documentation, examples or class definitions

_No response_

### Pseudocode or screenshots

_No response_
