### 2022-12-10 Artelia/NIMPHS#14 @Failxxx
Thank you very much for your recommendations. I find your idea relevant in this case because indeed, a lot of work has to be done by the user before being able to use the add-on. I have been dealing with this issue for a long time, even for other add-ons I have made, but never thought the it could be possible to proceed like you suggest. I will definitely give it a try!

I will also add what you are pointing out @bnase. This will motivate me to deal with #12 haha.
