### 2022-12-10 Artelia/NIMPHS#15 @Failxxx
Do you have any links or documentation for macFoam? I have googled it but I cannot find anything. Then I will probably be able to give you a better answer for your second question. Thank you.
