### 2022-12-10 Artelia/NIMPHS#15 @bnase
https://openfoam.org/version/macos/
