### 2022-12-23 #4868 @Failxxx
> sorry for the delay @Failxxx . You mentioned 2w ago that you would work on the issues raised, but have you fixed those? I don't see any commits.

No problem. Indeed I did not have the time to work on those yet. Is it mandatory for now? I don't think I will have the time by the end of the next week.
