### 2022-12-28 Artelia/NIMPHS#15 @Failxxx
> [openfoam.org/version/macos](https://openfoam.org/version/macos/)

If I understand correctly, the link you provided gives information on how to install and use OpenFOAM on macOS. However, this add-on does not use OpenFOAM or any CFD software to work. It can only post-process data computed through OpenFOAM or TELEMAC.

Is it because the installation instructions for macOS are not available for the moment that you propose to have this add-on work with macFOAM (which I suppose is OpenFOAM but for macOS?)?
