+++ .github/workflows/docs.yml
@@ -2,12 +2,16 @@ name: Build documentation
 on:
   pull_request: # Run in pull requests
 
-  workflow_dispatch: # Allow to run this workflow manually
+  workflow_dispatch: # Variables available when the workflow is manually triggered
     inputs:
       cache:
         description: "Use build cache"
         required: false
         default: "true"
+      publish:
+        description: "Publish documentation"
+        required: false
+        default: "false"
 
   push:
     tags: # Run when a new release is created
@@ -17,7 +21,7 @@ on:
 
 env:
   USE_CACHE: ${{ (github.event_name == 'workflow_dispatch' && github.event.inputs.cache == 'true') || (github.event_name == 'pull_request') }}
-  DEPLOY: ${{ github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags') }}
+  DEPLOY: ${{ github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags') || github.event.inputs.publish == 'true' }}
 
 jobs:
   build:
@@ -96,8 +100,8 @@ jobs:
         if: github.event_name == 'push' && github.ref == 'refs/heads/main'
         run: echo "Deploy dev documentation, not implemented yet"
 
-      - name: Deploy on release
-        if: startsWith(github.ref, 'refs/tags')
+      - name: Deploy on release or manually
+        if: startsWith(github.ref, 'refs/tags') || github.event.inputs.deploy == 'true'
         uses: peaceiris/actions-gh-pages@v3.7.3
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
