+++ .github/workflows/docs.yml
@@ -28,9 +28,9 @@ jobs:
     name: Build documentation
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2.4.2
+      - uses: actions/checkout@v2.6.0
 
-      - uses: actions/setup-python@v4.0.0
+      - uses: actions/setup-python@v4.4.0
         with:
           python-version: "3.10"
 
@@ -49,7 +49,7 @@ jobs:
         run: pip install -r requirements.txt
 
       - name: Cache docs _build directory
-        uses: actions/cache@v3.0.4
+        uses: actions/cache@v3.2.2
         if: env.USE_CACHE == 'true'
         with:
           path: docs/_build/
@@ -86,7 +86,7 @@ jobs:
     needs: build
     if: github.ref == 'refs/heads/main' || startsWith(github.ref, 'refs/tags')
     steps:
-      - uses: actions/checkout@v2.4.2
+      - uses: actions/checkout@v2.6.0
 
       - uses: actions/download-artifact@v3.0.0
         with:
