### 2023-01-25 #4868 @Failxxx
Hello. Here is an update :

- I merged your suggestions @hugoledoux in https://github.com/Artelia/NIMPHS/pull/16

- We are currently discussing about the following issue opened by @liberostelios: https://github.com/Artelia/NIMPHS/issues/15. I try to really understand what the issue is.

- The last issue will need more work. However, since this it is about an improvement, the add-on can still be installed and used normally. I will only be able to add suggested information in the installation process soon. I am working on a solution to better the process but this is not as easy as I thought haha. I will share my ideas and findings in the issue.
