+++ nimphs/__init__.py
@@ -7,8 +7,12 @@ from bpy.app.handlers import frame_change_pre, frame_change_post, save_pre
 from bpy.types import Scene, Object, TOPBAR_MT_file_import, VIEW3D_MT_editor_menus, VIEW3D_HT_tool_header
 
 import os
+import json
+from pathlib import Path
 
 from . import auto_load
+
+LOAD_INSTALLER = None
 
 bl_info = {
     "name": "NIMPHS",
@@ -24,12 +28,6 @@ bl_info = {
     "tracker_url": "https://github.com/Artelia/NIMPHS/issues"
 }
 
-# Test environment, inspired from:
-# https://github.com/JacquesLucke/animation_nodes/blob/master/animation_nodes/__init__.py
-# current_directory = os.path.dirname(os.path.abspath(__file__))
-# addons_directory = os.path.dirname(current_directory)
-# addon_name = os.path.basename(current_directory)
-
 if bl_version is not None and bl_version < (3, 0, 0):
     message = ("\n\n"
                "NIMPHS requires at least Blender 3.0.\n"
@@ -37,104 +35,107 @@ if bl_version is not None and bl_version < (3, 0, 0):
                "Please download the latest official release.")
     raise Exception(message)
 
-try:
-    import pyvista
-except BaseException:
-    pass
-
-if "pyvista" not in globals():
-    message = ("\n\n"
-               "NIMPHS depends on the pyvista library.\n"
-               "Unfortunately the Blender built you are using does not have this library.\n"
-               "Please checkout the documentation to fix this issue (installation guide).")
-    raise Exception(message)
-
-# Register
-from nimphs.menus.menus import nimphs_menus_draw
-from nimphs.properties.shared.nimphs_scene import NIMPHS_Scene
-from nimphs.properties.shared.nimphs_object import NIMPHS_Object
-from nimphs.operators.telemac.telemac_import_file import import_telemac_menu_draw
-from nimphs.operators.openfoam.openfoam_import_file import import_openfoam_menu_draw
-from nimphs.properties.utils.others import register_custom_progress_bar, nimphs_on_save_pre
-from nimphs.operators.utils.sequence import (
-    update_openfoam_streaming_sequences,
-    update_telemac_streaming_sequences,
-    update_telemac_mesh_sequences,
-)
+# Load add-on state information
+file_path = os.path.join(os.path.abspath(Path(__file__).parent), "nimphs.json")
+with open(file_path, "r+", encoding='utf-8') as file:
+    data = json.load(file)
+    if data["installation"]["state"] == 'DONE':
+        LOAD_INSTALLER = False
+    elif data["installation"]["state"] == 'INSTALL':
+        LOAD_INSTALLER = True
 
 # Setup logger
 import logging
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
+# Proceed to normal imports (installation is done)
+if LOAD_INSTALLER is False:
+    from nimphs.menus.menus import nimphs_menus_draw
+    from nimphs.properties.shared.nimphs_scene import NIMPHS_Scene
+    from nimphs.properties.shared.nimphs_object import NIMPHS_Object
+    from nimphs.operators.telemac.telemac_import_file import import_telemac_menu_draw
+    from nimphs.operators.openfoam.openfoam_import_file import import_openfoam_menu_draw
+    from nimphs.properties.utils.others import register_custom_progress_bar, nimphs_on_save_pre
+    from nimphs.operators.utils.sequence import (
+        update_openfoam_streaming_sequences,
+        update_telemac_streaming_sequences,
[…]
