### 2023-02-05 e11034b
[ADD-ON] Added installer panel + properties
