### 2023-02-05 254e1b7
[ADD-ON] Added basic installer operator
