### 2023-02-05 7a7dbba
[ADD-ON] Now saving state of install after successful installation
