### 2023-02-05 046cd0b
[INSTALL] Added ability to reset installer state from add-on preferences
