### 2023-02-05 ea4cb3f
[INSTALL] Fixed minor issues
