### 2023-02-05 2c8057d
[INSTALL] Added 'force' setting in installer
