### 2023-02-05 dbfc956
[Docs] Modified instructions for the new installation process
