### 2023-02-05 #4868 @Failxxx
Hello @hugoledoux @liberostelios @bnase

I finally had the time to improve the installation process. It is now published in version 0.4.2.
See associated PR: https://github.com/Artelia/NIMPHS/pull/17

Could you test and validate it please?
Thank you very much and sorry for the delay...
