+++ nimphs/requirements.txt
@@ -1,5 +1,4 @@
 # Requirements for the 'CLASSIC' configuration.
 
-pip
 pyvista >= 0.35.1
 matplotlib
