### 2023-02-11 b73f778
[Install] Added error message in case of installation failure
