### 2023-02-11 2b99038
[Docs] Added information in installation instructions
