+++ docs/source/spelling_wordlist.txt
@@ -24,3 +24,4 @@ Voxels
 voxels
 vdb
 VDB
+pyopenvdb
