+++ nimphs/__init__.py
@@ -25,7 +25,7 @@ bl_info = {
     "name": "NIMPHS",
     "description": "Numerous Instruments to Manipulate and Post-process Hydraulic Simulations",
     "author": "Félix Olart, Thibault Oudart",
-    "version": (0, 4, 2),
+    "version": (0, 4, 3),
     "blender": (3, 0, 0),
     "location": "File > Import",
     "warning": "This version is still in development.",
