### 2023-03-03 #4868 @Failxxx
Here are the requested information:
* Tagged release: https://github.com/Artelia/NIMPHS/releases/tag/v0.4.3.
* Archived software on Zenodo: https://zenodo.org/record/7696769.
* I confirm that the archival deposit has the right metadata.
* DOI of the archived version: 10.5281/zenodo.7696769

Thank you!
